Fix: Refactor (#9)
* Refactors episode details presentation

Removes the dedicated EpisodeDetailsReducer and integrates episode details presentation directly into the CharacterDetailsReducer.
This simplifies the state management and allows for a more streamlined flow when displaying episode information.

Updates the CharacterDetailsView to use a sheet for displaying episode details, triggered by selecting an episode from the list.

* Increments marketing version to 1.0.2

Updates the marketing version number in the project settings.

This change reflects a minor update to the application,
and ensures that the correct version number is displayed
in the App Store and other distribution channels.